src/reconciler/requirements.txt
boto3>=1.34.0
PyGithub>=2.1.0
